requirements.txt
astroid==2.2.5
bootstrap4==0.1.0
colorama==0.4.1
Django==2.2.1
django-background-tasks==1.2.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-compat==1.0.15
django-crispy-forms==1.7.2
django-material==1.5.6
django-simple-search==1.0.2
django-smart-selects==1.5.4
django-watson==1.5.2
entrypoints==0.3
flake8==3.7.7
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
